refactor: move Transformer class to middlewares folder
