+++ tests/integration/polygonParts/history.spec.ts
@@ -10,9 +10,9 @@ import { getApp } from '../../../src/app';
 import { ConnectionManager } from '../../../src/common/connectionManager';
 import { SERVICES } from '../../../src/common/constants';
 import { ApplicationConfig, DbConfig } from '../../../src/common/interfaces';
-import { Transformer } from '../../../src/common/middlewares/transformer';
 import { createConnectionOptions } from '../../../src/common/utils';
-import { ValidationEntityQuery, ValidatePolygonPartsRequestBody } from '../../../src/polygonParts/controllers/interfaces';
+import { Transformer } from '../../../src/middlewares/transformer';
+import { ValidatePolygonPartsRequestBody, ValidationEntityQuery } from '../../../src/polygonParts/controllers/interfaces';
 import { History } from '../../../src/polygonParts/DAL/history';
 import { PolygonPart } from '../../../src/polygonParts/DAL/polygonPart';
 import { namingStrategy } from '../../../src/polygonParts/DAL/utils';
