+++ tests/integration/polygonParts/intersection.spec.ts
@@ -15,8 +15,8 @@ import { getApp } from '../../../src/app';
 import { ConnectionManager } from '../../../src/common/connectionManager';
 import { SERVICES } from '../../../src/common/constants';
 import { ApplicationConfig, DbConfig } from '../../../src/common/interfaces';
-import { Transformer } from '../../../src/common/middlewares/transformer';
 import { createConnectionOptions } from '../../../src/common/utils';
+import { Transformer } from '../../../src/middlewares/transformer';
 import { History } from '../../../src/polygonParts/DAL/history';
 import { PolygonPart } from '../../../src/polygonParts/DAL/polygonPart';
 import { namingStrategy } from '../../../src/polygonParts/DAL/utils';
