+++ tests/integration/polygonParts/process.spec.ts
@@ -2,16 +2,16 @@ import jsLogger from '@map-colonies/js-logger';
 import { trace } from '@opentelemetry/api';
 import { multiPolygon, polygon } from '@turf/helpers';
 import config from 'config';
+import { Feature, Polygon } from 'geojson';
 import { StatusCodes as httpStatusCodes } from 'http-status-codes';
 import { container } from 'tsyringe';
 import { DataSourceOptions } from 'typeorm';
-import { Feature, Polygon } from 'geojson';
 import { getApp } from '../../../src/app';
 import { ConnectionManager } from '../../../src/common/connectionManager';
 import { SERVICES } from '../../../src/common/constants';
 import { DbConfig } from '../../../src/common/interfaces';
-import { Transformer } from '../../../src/common/middlewares/transformer';
 import { createConnectionOptions } from '../../../src/common/utils';
+import { Transformer } from '../../../src/middlewares/transformer';
 import { ProcessPolygonPartsRequestBody } from '../../../src/polygonParts/controllers/interfaces';
 import { History } from '../../../src/polygonParts/DAL/history';
 import { PolygonPart } from '../../../src/polygonParts/DAL/polygonPart';
